+++ src/TAGCompound.java
@@ -68,7 +68,7 @@ public class TAGCompound extends TAGComponent{
     }
 
     @Override
-    public int getTypeId() {
+    public byte getTypeId() {
         return 10;
     }
