+++ src/TAGInt.java
@@ -33,7 +33,7 @@ public class TAGInt extends TAGComponent{
 	}
 
 	@Override
-	public int getTypeId() {
+	public byte getTypeId() {
 		return 3;
 	}
