+++ src/TAGLong.java
@@ -34,7 +34,7 @@ public class TAGLong extends TAGComponent{
 
 
 	@Override
-	public int getTypeId() {
+	public byte getTypeId() {
 		return 4;
 	}
